Sample PDF support added
