--- src/middlewares/pdfTableofContents/config.ts
@@ -1,10 +1,0 @@
-import path from 'path';
-
-export = {
-  fontPath: path.join(__dirname, 'src', 'data', 'monkey.ttf'),
-  leftMarginEnd: 78,
-  horizontalMax: 595,
-  verticalMax: 842,
-  verticalMargin: 100,
-  lineSpacing: 20,
-};
